Refactor: Consolidate user-defined type tests and remove redundant file
- Moved tests from `relvar-core/src/types/user_defined_test.rs` to `relvar-core/src/values/scalar.rs`.
- Deleted `relvar-core/src/types/user_defined_test.rs`.
- Removed module declaration from `relvar-core/src/types/mod.rs`.

This simplifies the project structure by co-locating tests with the code they test and removing an anomalous test file from the source tree.
