+++ relvar-core/src/types/mod.rs
@@ -45,8 +45,6 @@
 pub mod relation_type;
 pub mod scalar;
 pub mod tuple_type;
-#[cfg(test)]
-mod user_defined_test;
 
 pub use relation_type::RelationType;
 pub use scalar::ScalarType;
